refactor: Adjust structure
