+++ src/components/App.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
-import CategoryPage from "./CategoryPage";
+import CategoryPage from "./Category/CategoryPage";
 
 const App = () => {
   return (
